refact: Move ActivateScopes support under context.internal
