+++ core/src/main/java/io/github/cdiunit/core/context/internal/CdiContext.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.cdiunit.internal.activatescopes;
+package io.github.cdiunit.core.context.internal;
 
 import java.lang.annotation.Annotation;
 import java.util.HashMap;
