+++ core/src/main/java/io/github/cdiunit/internal/DefaultDiscoveryContext.java
@@ -31,8 +31,8 @@ import org.jboss.weld.environment.se.Weld;
 import org.jboss.weld.environment.se.WeldSEBeanRegistrant;
 
 import io.github.cdiunit.ProducesAlternative;
-import io.github.cdiunit.core.context.internal.ContextTrackerExtension;
-import io.github.cdiunit.internal.activatescopes.ScopesExtension;
+import io.github.cdiunit.core.context.internal.ContextActivator;
+import io.github.cdiunit.core.context.internal.ContextTracker;
 
 class DefaultDiscoveryContext implements DiscoveryExtension.Context {
 
@@ -246,8 +246,8 @@ class DefaultDiscoveryContext implements DiscoveryExtension.Context {
         decorators.forEach(weld::addDecorator);
         interceptors.forEach(weld::addInterceptor);
 
-        weld.addExtension(new ScopesExtension(scopeTypes));
-        weld.addExtension(new ContextTrackerExtension(scopeTypes));
+        weld.addExtension(new ContextActivator(scopeTypes));
+        weld.addExtension(new ContextTracker(scopeTypes));
     }
 
 }
