+++ spock/src/main/java/io/github/cdiunit/spock/internal/ActivateScopes.java
@@ -22,9 +22,9 @@ import org.spockframework.runtime.extension.IMethodInvocation;
 import org.spockframework.runtime.extension.IStore;
 
 import io.github.cdiunit.IsolationLevel;
+import io.github.cdiunit.core.context.Scopes;
 import io.github.cdiunit.internal.TestLifecycle;
 import io.github.cdiunit.internal.TestMethodHolder;
-import io.github.cdiunit.internal.activatescopes.Scopes;
 
 public class ActivateScopes implements IMethodInterceptor {
